Resolve MacOS docs (#168)
* Rename page from "Mac OSX" to "macOS"

Mac OS X was officially renamed to macOS back in 2016, seven years ago.
For now, we'll leave the filename alone.

* Remove information that is not actionable

There are three bullets here that don't provide guidance
on what to do.  There is no indication of how this information
can be used, so it just confuses matters.  I'd like to keep the
info focused on steps needed to set up the dev box.

* Clarify steps in installing needed tools and forking repo

* Always run `cd` as a separate step

* Don't combine it with the next command via `&&`.
  If the second command fails and we re-run the line,
  we'll be in the wrong directory.
* Explicitly `cd` back out of the subdirectory so we are back in the repo root

Also:
* Format the Troubleshooting sections the same as the main sections,
  that is, with bullets.

* Remove Troubleshooting for yarn/fibers "code 127"

Error code 127 is "command not found".

The executable for Node.js is traditionally named `node`, however there exists
another software package out there that also uses that name.  To avoid a conflict,
**Ubuntu** installs the Node.js executable as `nodejs` even though that breaks
scripts that expect a `node` command to be available.

To support Ubuntu, the fibers package.json calls it as a fallback.
Instead of checking for the presence of `node` and `nodejs` binaries
and calling the appropriate one, it blindly calls `node` and assumes
that any failure _must_ be caused by that command's absence.

From fibers' package.json:

      ```json
      "scripts": {
        "install": "node build.js || nodejs build.js",
        "test": "node test.js || nodejs test.js"
      },
      ```

While this is a fault in the fibers install script, our troubleshooting
workarounds should focus on why the `node build.js` is failing, and not
add additional confusing steps to deal with secondary failures.  That is,
once the user follows the instructions in Troubleshooting §2, then §1
becomes moot.  Let's minimize the number of steps the user has to do.
Delete §1 entirely.

[The "§2" and "§3" below refer to the _original_ section numbers, before deleting §1.]

===

After §2 is followed, the build _still_ fails.  There is nothing in the error
message that indicates "bcrypt" (see §3), e.g. the error message listed at
the top of §3.  Nonetheless, Debdut instructed me (in a private conversation)
to follow §3 anyway, and indeed my next build attempt succeeded.

Add instructions to §2 _explicitly_ guiding the user to follow §3.
Leave §3 in place in case the user encounters its error message and
does _not_ encounter §2's error.

* Restore deleted instruction to run `yarn`

A recent update accidentally deleted a crucial step: Running `yarn`

Also:
* Avoid comments in blocks of shell commands, since `zsh` (by default)
  doesn't recognize `#` as a comment character.  So split up the
  `yarn dev` vs `dsv` instructions into multiple blocks.
* Move the reference to the Troubleshooting section up to the top,
  so we know that it exists right from the beginning.

* (minor) Fix Markdown formatting

* Update mac-osx.md

---------

Co-authored-by: Noach Magedman <noach@seekingalpha.com>